Fix mixed precision training attempt 2

+++ train.py
@@ -129,7 +129,7 @@ for epoch in range(int(config["hyperparameters"]["NUM_EPOCHS"])):
             optimizer.step()
 
         train_loss.update(val=loss.item(), n=images.size(0))
-        segmaps = segmaps.to('cpu')
+        segmaps = segmaps.to('cpu').to(torch.float32)
         segmaps_classid = segmaps.argmax(1)
         train_iou.add(pred_segmaps.detach(), segmaps_classid.detach())
         train_pbar.update(1)
@@ -171,11 +171,11 @@ for epoch in range(int(config["hyperparameters"]["NUM_EPOCHS"])):
             val_loss_val /= 2
             val_loss.update(val=val_loss_val.item(), n=2*prelabels[0].size(0))
 
-            prelabels[0] = prelabels[0].to('cpu')
-            postlabels[0] = postlabels[0].to('cpu')
+            prelabels[0] = prelabels[0].to('cpu').to(torch.float32)
+            postlabels[0] = postlabels[0].to('cpu').to(torch.float32)
 
-            pre_preds = pre_preds.to('cpu')
-            post_preds = post_preds.to('cpu')
+            pre_preds = pre_preds.to('cpu').to(torch.float32)
+            post_preds = post_preds.to('cpu').to(torch.float32)
 
             pre_gt_classid = prelabels[0].argmax(1)
             post_gt_classid = postlabels[0].argmax(1)
